+++ src/bw2calc/__init__.py
@@ -11,7 +11,7 @@ __all__ = [
     "FastScoresOnlyMultiLCA",
 ]
 
-__version__ = "2.3.2"
+__version__ = "2.4.0"
 
 
 import platform